backend/requirements.txt
fastapi
uvicorn[standard]
authlib
bcrypt
pandas
numpy
python-dotenv
pydantic
